Update jisho downloader.py
Changed the code that searches the kanji in jisho's api. It wasn't searching the kanji in quotes so when searching for a kanji that's not in the dictionary, instead of giving you and error and continuing with the next one, it was outputing whatever jisho considered to be most similar to what you were searching, this can lead to big mistakes.